patterns: Make offset a size_t instead of off_t
This allows comparing against sentence.size()

+++ apt-pkg/cachefilter-patterns.h
@@ -71,7 +71,7 @@ struct PatternTreeParser
 
    struct State
    {
-      off_t offset = 0;
+      size_t offset = 0;
    };
 
    APT::StringView sentence;
